backend/reecon/requirements.txt
Django==5.1.7
django-constance[redis]==4.3.2
django-redis==5.4.0
django-rq==3.0.0
djangorestframework==3.16.0
hiredis==3.1.0
markdown==3.7
openai==1.69.0
praw==7.8.1
psycopg[c]==3.2.6
pydantic==2.11.1
python-decouple==3.8
tenacity==9.0.0
tiktoken==0.9.0
